refactor: Platforms types
